+++ src/main/java/de/maxhenkel/peek/interfaces/PeekDecoratedPot.java
@@ -1,0 +1,9 @@
+package de.maxhenkel.peek.interfaces;
+
+import net.minecraft.world.item.ItemStack;
+
+public interface PeekDecoratedPot {
+
+    ItemStack peek$getContainedItem();
+
+}