+++ src/main/java/de/maxhenkel/peek/mixin/DecoratedPotBlockEntityMixin.java
@@ -1,0 +1,54 @@
+package de.maxhenkel.peek.mixin;
+
+import de.maxhenkel.peek.Peek;
+import de.maxhenkel.peek.interfaces.PeekDecoratedPot;
+import net.minecraft.core.BlockPos;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.protocol.Packet;
+import net.minecraft.network.protocol.game.ClientGamePacketListener;
+import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.entity.DecoratedPotBlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+
+import javax.annotation.Nullable;
+
+@Mixin(DecoratedPotBlockEntity.class)
+public class DecoratedPotBlockEntityMixin extends BlockEntity implements PeekDecoratedPot {
+
+    @Shadow
+    private ItemStack item;
+
+    public DecoratedPotBlockEntityMixin(BlockEntityType<?> blockEntityType, BlockPos blockPos, BlockState blockState) {
+        super(blockEntityType, blockPos, blockState);
+    }
+
+    @Override
+    public ItemStack peek$getContainedItem() {
+        return item;
+    }
+
+    @Nullable
+    @Override
+    public Packet<ClientGamePacketListener> getUpdatePacket() {
+        if (!Peek.CONFIG.sendDecoratedPotDataToClient.get()) {
+            return super.getUpdatePacket();
+        }
+        return ClientboundBlockEntityDataPacket.create(this);
+    }
+
+    @Override
+    public CompoundTag getUpdateTag() {
+        if (!Peek.CONFIG.sendDecoratedPotDataToClient.get()) {
+            return super.getUpdateTag();
+        }
+        CompoundTag tag = new CompoundTag();
+        saveAdditional(tag);
+        return tag;
+    }
+
+}