+++ src/main/java/de/maxhenkel/peek/mixin/ItemRendererMixin.java
@@ -1,6 +1,7 @@
 package de.maxhenkel.peek.mixin;
 
 import com.mojang.blaze3d.vertex.PoseStack;
+import de.maxhenkel.peek.events.DecoratedPotRenderEvents;
 import de.maxhenkel.peek.events.ShulkerRenderEvents;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.ItemRenderer;
@@ -17,7 +18,8 @@ public class ItemRendererMixin {
 
     @Inject(method = "render", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/block/model/ItemTransform;apply(ZLcom/mojang/blaze3d/vertex/PoseStack;)V", shift = At.Shift.AFTER))
     private void apply(ItemStack itemStack, ItemDisplayContext itemDisplayContext, boolean bl, PoseStack poseStack, MultiBufferSource multiBufferSource, int i, int j, BakedModel bakedModel, CallbackInfo ci) {
-        if (!ShulkerRenderEvents.isShulkerRenderStack(itemStack)) {
+        //TODO Create a common method to detect custom rendered items
+        if (!ShulkerRenderEvents.isShulkerRenderStack(itemStack) && !DecoratedPotRenderEvents.isPotRenderStack(itemStack)) {
             return;
         }
         poseStack.popPose();