+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed items being flipped on shulker boxes
+- Added decorated pot hints